requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
boto3==1.34.144
python-dotenv==1.0.0
PyPDF2==3.0.1
pdf2image==1.17.0
pillow==10.0.1
requests==2.31.0
aiofiles==23.2.0
jinja2==3.1.2
python-jose[cryptography]==3.3.0